app/v2.0/requirements.txt
PyQt5==5.15.7
pyserial==3.5
requests==2.27.1
pyinstaller==5.7.0